+++ Sources/KeyboardKitSwiftUI/_Deprecations/KeyboardShiftState+Deprecations.swift
@@ -6,7 +6,7 @@ public extension KeyboardShiftState {
     @available(*, deprecated, renamed: "standardButtonImage")
     var systemImage: Image { standardButtonImage }
     
-    @available(*, deprecated, renamed: "standardButtonImage")
+    @available(*, deprecated, message: "")
     func systemKeyboardButtonImage(for context: KeyboardContext) -> Image {
         switch context.keyboardType {
         case .alphabetic(let state): return state.standardButtonImage